homework_07_streaming/pyproject.toml
[project]
name = "homework-07-streaming"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kafka-python>=2.3.0",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
]
